+++ CameraSample/CameraSample/Content/FilterBanner/FilterBanner.swift
@@ -7,6 +7,7 @@
 //
 
 import SwiftUI
+import Combine
 
 struct FilterBannerView: View {
     @Binding var isShowBanner: Bool
@@ -73,6 +74,38 @@ struct FilterBannerView: View {
         }
     }
 }
+
+final class FilterBannerViewModel: ObservableObject {
+    @Published var selectedFilter: FilterType?
+    private var cancelables: [Cancellable] = []
+    var isSelectedPixellate: Bool {
+        return selectedFilter == .pixellate
+    }
+
+    var isSelectedSepiaTone: Bool {
+        return selectedFilter == .sepiaTone
+    }
+
+    var isSelectedSharpenLuminance: Bool {
+        return selectedFilter == .sharpenLuminance
+    }
+
+    var isSelectedPhotoEffectMono: Bool {
+        return selectedFilter == .photoEffectMono
+    }
+
+    var isSelectedGaussianBlur: Bool {
+        return selectedFilter == .gaussianBlur
+    }
+
+    init() {
+        let subscriber = $selectedFilter.sink { (type) in
+            print("\(type.debugDescription)")
+        }
+        cancelables.append(subscriber)
+    }
+}
+
 struct FilterPreviewView_Previews: PreviewProvider {
     static var previews: some View {
         FilterBannerView(isShowBanner: .constant(true), selectedFilterType: .constant(.gaussianBlur), uiImage: .constant(UIImage(named: "snap")))
